Migrate top panel
